Move parse_fastani_file under methods/
