+++ pyani_plus/private_cli.py
@@ -33,7 +33,7 @@ import typer
 from rich.progress import track
 
 from pyani_plus import db_orm, tools
-from pyani_plus.methods import method_anib
+from pyani_plus.methods import method_anib, method_fastani
 from pyani_plus.utils import file_md5sum
 
 app = typer.Typer()
@@ -257,38 +257,6 @@ def log_comparison(  # noqa: PLR0913
     return 0
 
 
-def parse_fastani_file(filename: Path) -> tuple[Path, Path, float, int, int]:
-    """Parse a single-line fastANI output file extracting key fields as a tuple.
-
-    Return (ref genome, query genome, ANI estimate, orthologous matches,
-    sequence fragments) tuple.
-
-    :param filename: Path, path to the input file
-
-    Extracts the ANI estimate (which we return in the range 0 to 1), the
-    number of orthologous matches (int), and the number of sequence
-    fragments considered from the fastANI output file (int).
-
-    We assume that all fastANI comparisons are pairwise: one query and
-    one reference file. The fastANI file should contain a single line.
-
-    fastANI *can* produce multi-line output, if a list of query/reference
-    files is given to it.
-    """
-    with filename.open() as handle:
-        line = handle.readline().strip().split()
-    if not line:  # No file content; either run failed or no detectable similarity
-        msg = f"Input file {filename} is empty"
-        raise ValueError(msg)
-    return (
-        Path(line[0]),
-        Path(line[1]),
-        0.01 * float(line[2]),
-        int(line[3]),
-        int(line[4]),
-    )
-
-
 # Ought we switch the command line arguments here to match fastANI naming?
 # Note this omits maxmatch
 @app.command()
@@ -318,7 +286,7 @@ def log_fastani(  # noqa: PLR0913
     fastani_tool = tools.get_fastani()
 
     used_query, used_subject, identity, orthologous_matches, fragments = (
-        parse_fastani_file(fastani)
+        method_fastani.parse_fastani_file(fastani)
     )
     # Allowing for some variation in the filename paths here... should we?
     if used_query.stem != query_fasta.stem:
